Set preferred content size of RichTextViewController
Whenever the layout and the size of the text view within the RichTextViewController changes, we want to set the preferred content size of the RichTextViewController to reflect the content size fo the the text view. This allows us to determine the size of the view's content to be used in scenarios where we use the RichTextViewController as a child view of another container view controller.